Primary Templates separates
Templates separated from QuickScript.exe into the file system under
Templates sub folder. Only one is supported right this second. More to
follow quickly.
