+++ MetX/MetX/Data/XlgQuickScript.cs
@@ -15,10 +15,16 @@ namespace MetX.Data
 {
     public class CaseFreeDictionary : Dictionary<string, string>
     {
-        public new string this[string name]
+        public string this[string name]
         {
             get { return base[name.ToLower()]; }
             set { base[name.ToLower()] = value; }
+        }
+
+        public string View(string name)
+        {
+            name = name.ToLower();
+            return base.ContainsKey(name) ? base[name] : null;
         }
     }
 
@@ -52,16 +58,27 @@ namespace MetX.Data
     {
         private static string m_IndependentTemplate;
         private static string m_Template;
-        [XmlAttribute] public QuickScriptDestination Destination;
-        [XmlAttribute] public string DestinationFilePath;
-        [XmlAttribute] public string DiceAt;
-        [XmlAttribute] public Guid Id;
-        [XmlAttribute] public string Input;
-        [XmlAttribute] public string InputFilePath;
-        [XmlAttribute] public string Name;
-        [XmlAttribute] public string Script;
-        [XmlAttribute] public string SliceAt;
-        [XmlAttribute] public string Template;
+        [XmlAttribute]
+        public QuickScriptDestination Destination;
+        [XmlAttribute]
+        public string DestinationFilePath;
+        [XmlAttribute]
+        public string DiceAt;
+        [XmlAttribute]
+        public Guid Id;
+        [XmlAttribute]
+        public string Input;
+        [XmlAttribute]
+        public string InputFilePath;
+        [XmlAttribute]
+        public string Name;
+        [XmlAttribute]
+        public string Script;
+        [XmlAttribute]
+        public string SliceAt;
+        [XmlAttribute]
+        public string Template;
+
         public XlgQuickScript() { }
 
         public XlgQuickScript(string name, string script = null)
@@ -87,7 +104,7 @@ namespace MetX.Data
             {
                 compilerParameters.MainClass = "Processor.Program";
                 compilerParameters.OutputAssembly =
-                    Guid.NewGuid().ToString().Replace(new[] {"{", "}", "-"}, "").ToLower() + ".exe";
+                    Guid.NewGuid().ToString().Replace(new[] { "{", "}", "-" }, "").ToLower() + ".exe";
             }
 
             compilerParameters
@@ -103,9 +120,9 @@ namespace MetX.Data
             return compilerResults;
         }
 
-        public string ToCSharp(bool independent)
+        public string ToCSharp(XlgQuickScriptTemplateList templates, bool independent)
         {
-            string code = new GenInstance(this, independent).CSharp;
+            string code = new GenInstance(this, templates[Template], independent).CSharp;
             if (code.IsNullOrEmpty())
             {
                 return code;
@@ -142,7 +159,7 @@ namespace MetX.Data
             bool flag;
             if (commandArgs.Restore)
             {
-                logger.LogMessage(LogLevel.Verbose, "Restoring {0}...", (object) commandArgs.Input);
+                logger.LogMessage(LogLevel.Verbose, "Restoring {0}...", (object)commandArgs.Input);
                 string fileNameKey = BackupUtilities.CreateFileNameKey(commandArgs.Input);
                 try
                 {
@@ -167,7 +184,7 @@ namespace MetX.Data
                 flag = new FileArranger(commandArgs.Configuration, logger).Arrange(commandArgs.Input, commandArgs.Output, commandArgs.Backup);
                 if (!flag)
                 {
-                    logger.LogMessage(LogLevel.Error, "Unable to arrange {0}.", (object) commandArgs.Input);
+                    logger.LogMessage(LogLevel.Error, "Unable to arrange {0}.", (object)commandArgs.Input);
                 }
                 else
                 {
@@ -385,162 +402,6 @@ namespace MetX.Data
         {
             Name = name;
             Indent = 12;
-        }
-    }
-
-    public class GenInstance : List<GenArea>
-    {
-        public XlgQuickScriptTemplate Template;
-        public string CSharp
-        {
-            get
-            {
-                string originalArea = string.Empty;
-                foreach (string currScriptLine in m_Quick.Script.Lines())
-                {
-                    int indent = currScriptLine.Length - currScriptLine.Trim().Length;
-                    if (currScriptLine.Contains("~~Start:") || currScriptLine.Contains("~~Begin:"))
-                    {
-                        SetArea("Start");
-                    }
-                    else if (currScriptLine.Contains("~~Finish:") || currScriptLine.Contains("~~Final:") || currScriptLine.Contains("~~End:"))
-                    {
-                        SetArea("Finish");
-                    }
-                    else if (currScriptLine.Contains("~~ReadInput:") || currScriptLine.Contains("~~Read:") || currScriptLine.Contains("~~Input:"))
-                    {
-                        SetArea("ReadInput");
-                    }
-                    else if (currScriptLine.Contains("~~ClassMember:") || currScriptLine.Contains("~~ClassMembers:") || currScriptLine.Contains("~~Fields:") || currScriptLine.Contains("~~Field:")
-                             || currScriptLine.Contains("~~Members:") || currScriptLine.Contains("~~Member:"))
-                    {
-                        SetArea("ClassMembers");
-                    }
-                    else if (currScriptLine.Contains("~~ProcessLine:") || currScriptLine.Contains("~~ProcessLines:") || currScriptLine.Contains("~~Body:"))
-                    {
-                        SetArea("ProcessLine");
-                    }
-                    else if (currScriptLine.Contains("~~BeginString:"))
-                    {
-                        string stringName = currScriptLine.TokenAt(2, "~~BeginString:").Trim();
-                        if (string.IsNullOrEmpty(stringName))
-                        {
-                            continue;
-                        }
-                        m_CurrArea.Lines.Add("string " + stringName + " = //~~String " + stringName + "~~//;");
-                        originalArea = m_CurrArea.Name;
-                        SetArea("String " + stringName);
-                    }
-                    else if (currScriptLine.Contains("~~EndString:"))
-                    {
-                        if (m_CurrArea.Lines.IsNullOrEmpty())
-                        {
-                            m_CurrArea.Lines.Add("string.Empty");
-                        }
-                        else if (!m_CurrArea.Lines.Any(x => x.Contains("\"")))
-                        {
-                            m_CurrArea.Lines[0] = "@\"" + m_CurrArea.Lines[0];
-                            m_CurrArea.Lines[m_CurrArea.Lines.Count - 1] += "\"";
-                        }
-                        else if (!m_CurrArea.Lines.Any(x => x.Contains("`")))
-                        {
-                            m_CurrArea.Lines.TransformAllNotEmpty((line, index) => line.Replace("\"", "``"));
-                            m_CurrArea.Lines[0] = "@\"" + m_CurrArea.Lines[0];
-                            m_CurrArea.Lines[m_CurrArea.Lines.Count - 1] += "\".Replace(\"``\",\"\\\"\")";
-                        }
-                        else if (!m_CurrArea.Lines.Any(x => x.Contains("`")))
-                        {
-                            m_CurrArea.Lines.TransformAllNotEmpty((line, index) => "\t\"" + line.Replace("\"", "\\\"") + "\" + Enviornment.NewLine;" + Environment.NewLine);
-                            m_CurrArea.Lines[0] = "@\"" + m_CurrArea.Lines[0];
-                            m_CurrArea.Lines[m_CurrArea.Lines.Count - 1] += "\".Replace(\"``\",\"\\\"\")";
-                        }
-                        SetArea(originalArea);
-                        originalArea = null;
-                    }
-                    else if (currScriptLine.Contains("~~:"))
-                    {
-                        if (m_CurrArea.Name == "ClassMembers")
-                        {
-                            m_CurrArea.Lines.Add(XlgQuickScript.ExpandScriptLineToSourceCode(currScriptLine, -1));
-                        }
-                        else
-                        {
-                            m_CurrArea.Lines.Add(XlgQuickScript.ExpandScriptLineToSourceCode(currScriptLine, indent));
-                        }
-                    }
-                    else if (originalArea != null)
-                    {
-                        m_CurrArea.Lines.Add(currScriptLine);
-                    }
-                    else
-                    {
-                        m_CurrArea.Lines.Add((new string(' ', indent + m_CurrArea.Indent)) + currScriptLine);
-                    }
-                }
-
-                StringBuilder sb = new StringBuilder(Template.Views[m_Independent ? "Exe" : "Native"]);
-                foreach (GenArea area in this)
-                {
-                    sb.Replace("//~~" + area.Name + "~~//", String.Join(Environment.NewLine, area.Lines));
-                }
-
-                if (m_Independent)
-                {
-                    sb.Replace("//~~InputFilePath~~//", "\"" + m_Quick.InputFilePath.LastToken(@"\") + "\"");
-                    sb.Replace("//~~Namespace~~//", (m_Quick.Name + "_" + DateTime.UtcNow.ToString("G") + "z").AsFilename());
-                    sb.Replace("//~~NameInstance~~//", m_Quick.Name + " at " + DateTime.Now.ToString("G"));
-
-                    switch (m_Quick.Destination)
-                    {
-                        case QuickScriptDestination.TextBox:
-                        case QuickScriptDestination.Clipboard:
-                        case QuickScriptDestination.Notepad:
-                            sb.Replace("//~~DestinationFilePath~~//", "Path.GetTempFileName()");
-                            break;
-
-                        case QuickScriptDestination.File:
-                            sb.Replace("//~~DestinationFilePath~~//", "\"" + m_Quick.DestinationFilePath.LastToken(@"\") + "\"");
-                            break;
-                    }
-                }
-
-                for (int i = 0; i < 10; i++)
-                {
-                    sb.Replace(Environment.NewLine + Environment.NewLine, Environment.NewLine);
-                }
-                return sb.ToString();
-            }
-        }
-
-        private readonly bool m_Independent;
-        private readonly XlgQuickScript m_Quick;
-        private GenArea m_CurrArea;
-
-        public GenInstance(XlgQuickScript quick, bool independent)
-        {
-            m_Quick = quick;
-            m_Independent = independent;
-
-            m_CurrArea = new GenArea("ProcessLine");
-            AddRange(new[]
-            {
-                m_CurrArea,
-                new GenArea("ClassMembers", 8),
-                new GenArea("Start"),
-                new GenArea("Finish"),
-                new GenArea("ReadInput"),
-            });
-        }
-
-        private void SetArea(string areaName)
-        {
-            foreach (GenArea area in this.Where(area => area.Name == areaName))
-            {
-                m_CurrArea = area;
-                return;
-            }
-            m_CurrArea = new GenArea(areaName);
-            Add(m_CurrArea);
         }
     }
 }
